+++ packages/server/src/auth/auth.controller.ts
@@ -1,12 +1,11 @@
 import { Controller, Post, UseGuards } from '@nestjs/common';
-import {ApiTags, ApiBody, ApiResponse} from '@nestjs/swagger';
+import { ApiTags, ApiBody, ApiResponse } from '@nestjs/swagger';
 import { AuthService } from './auth.service';
 import { LocalAuthGuard } from './local-auth.guard';
 import { JwtAuthGuard } from './jwt-auth.guard';
 import { UserSession } from '../users/users.session.decorator';
 import { User } from '../users/users.interface';
-import {LoginDto, LoginResponse} from './dto/login.dto';
-import {ResourcesResponse} from "../resources/resources.dto";
+import { LoginRequest, LoginResponse } from '@shared/auth';
 
 @ApiTags('auth')
 @Controller('auth')
@@ -15,7 +14,7 @@ export class AuthController {
 
     @UseGuards(LocalAuthGuard)
     @Post('login')
-    @ApiBody({ type: LoginDto })
+    @ApiBody({ type: LoginRequest })
     @ApiResponse({ status: 200, type: LoginResponse })
     login(@UserSession() user: User): LoginResponse {
         return this.authService.login(user);
