+++ packages/server/src/auth/auth.service.ts
@@ -2,13 +2,13 @@ import { Injectable } from '@nestjs/common';
 import { JwtService } from '@nestjs/jwt';
 import { UsersService } from '../users/users.service';
 import { User } from '../users/users.interface';
-import {LoginDto, LoginResponse} from './dto/login.dto';
+import { LoginRequest, LoginResponse } from '@shared/auth';
 
 @Injectable()
 export class AuthService {
     constructor(private readonly usersService: UsersService, private jwtService: JwtService) {}
 
-    validateUser({ email, password }: LoginDto) {
+    validateUser({ email, password }: LoginRequest) {
         const user = this.usersService.findOne(email);
         if (user) {
             const passwordMatch = user.password === password;
